+++ packages/create-mastra/CHANGELOG.md
@@ -1,5 +1,7 @@
 # create-mastra
 
+## 1.7.3-alpha.5
+
 ## 1.7.3-alpha.4
 
 ## 1.7.3-alpha.3